fix(viva): keep opening fail-closed and close before playback
Stop the 20s watchdog from replaying a cached caption onto the next audio
turn, destroy the viva window when committee playback never started, and
move dialog orchestration into useVoiceDefense with a testable start gate.
