+++ docs/bug_fix/voice-defense-ended-missing-close.md
@@ -1,0 +1,22 @@
+# 答辩总结页没有关闭按钮
+
+**状态**：已修复  
+**日期**：2026-08-14  
+**范围**：答辩间总结页 `EndedAct` 底栏
+
+## 现象
+
+保存转写后，底栏只剩「打开转写 / 生成评价 / 再来一场」。关闭只留在标题栏小叉上，录屏和浅色主题下几乎看不见。
+
+## 根因
+
+底栏把「关闭」当成无转写操作时的主按钮。一旦出现「保存转写」或「打开转写」，关闭从底栏消失，只剩图标按钮。
+
+## 修复
+
+有转写操作时，底栏末尾仍放带文案的「关闭」幽灵按钮。标题栏叉号补 Tooltip。
+
+## 验证
+
+- `pnpm exec vitest run test/voice-defense-ended-act.test.ts`
+- 保存转写后的总结页：底栏能看到「关闭」，点击后退出答辩间
