+++ docs/bug_fix/voice-defense-opening-ack-gate.md
@@ -183,3 +183,31 @@ ChatGPT Web Voice **没有 system 通道**。会前 brief 只能以 `relay_messa
 ## 边界（第六轮）
 
 字幕晚于音频是上游 ASR 的固有延迟，开场轮从 `speaking` 起同时放声和收字幕，二者保持同一轮。`stop_speaking` 截不断已经送出的前奏 PCM。没有宣布、只有提问时依赖 20 秒兜底或恢复触发，不会一直无声。
+
+> 第六轮的 `speaking` 提前放声和 20 秒盲放开策略已由下述第七轮替换。
+
+## 后续：20 秒盲兜底把上一轮字幕配到了下一轮声音
+
+打包版实机会话给出了可重复的证据链：15:36:07 连接成功，15:36:27 恰好在 20 秒看门狗到期时记录 `committee playback enabled reason=gate`；结束摘要为 `openingTriggerSent=false`，同一时段没有 `offline`、WebRTC `disconnected/failed` 或 DataChannel 错误。保存的转写依次出现：
+
+1. 「好的，我们来聚焦他们的核心主张。我想问……」；
+2. 「嗯。答辩现在开始。第一个问题：……」。
+
+这不是一次普通的网络卡顿。旧状态机把“时间到了”当作“已经听到开场”的证据，直接将门控设为 `open`；`applyOpening` 又把 `lastHiddenOpeningCaption` 回放到舞台，同时打开仍在继续的远端音频。因此屏幕显示的是缓存的上一轮文字，扬声器可能已经处于下一段声音，形成跨轮声画错配。之后宽松的宣布检测还会接受带「嗯」前缀的恢复轮。
+
+第七轮改为 fail-closed：
+
+1. 只有字幕从第一个字起命中「答辩现在开始」或 `The defense begins now` 才是合格开场；「嗯。答辩现在开始」「好的，答辩开始」仅用于识别异常/重启，不会开声、上台或开麦。
+2. 20 秒到期不再伪造 `heardOpening`，也不回放任何隐藏字幕；客户端无条件 `stop_speaking` 打断旧轮次，再发送要求精确首句的内部恢复指令。恢复后再超时则显示本地化错误，不把未知内容漏给用户。
+3. 删除恢复轮 `speaking` 一到就提前开声的旁路。远端音轨和 `<audio>` 只有在规范开场字幕命中后才启用，字幕与声音至少来自同一个委员消息。
+4. 通话计时不再由连接、上游 `speaking` 或先到字幕触发，而由 `onCommitteePlayback(true)`——即委员声音真正获准播放的时刻——触发且只触发一次。
+
+## 回归（第七轮）
+
+- `test/voice-defense-client.test.ts` 重放上述真实首轮：20 秒超时后仍为 `awaiting`、舞台无字幕、远端音轨关闭，并验证打断与精确恢复指令；后续「嗯。答辩现在开始」仍不能放声。
+- `test/voice-defense-protocol.test.ts` 验证超时本身永远不能证明开场已发生，且中英文规范前缀通过、带口癖前缀和宽松改写失败。
+- `test/voice-defense-timing.test.ts` 验证只有远端播放启用才起表，连接等待与重复播放通知不改变起点。
+
+## 边界（第七轮）
+
+远端声音使用 WebRTC RTP，流式字幕使用 DataChannel，当前私有协议没有提供两者共享的逐字时间戳；即使连接健康，抓包也观察到字幕增量间隔约 3.6 秒。网络抖动可能放大字幕的 token 级延迟，但不能解释本次“上一轮字幕 + 下一轮声音”的稳定错配。第七轮消除了跨消息错配；在不缓存并重放整段远端音频的前提下，不承诺卡拉 OK 式逐字同步。
