+++ mkdocs.yml
@@ -106,6 +106,8 @@ nav:
       - 语音答辩开场噪声字幕与重启: bug_fix/voice-defense-opening-noise-captions.md
       - 语音答辩开场回执把答辩重启: bug_fix/voice-defense-opening-ack-gate.md
       - 答辩总结页底栏按钮叠层: bug_fix/voice-defense-ended-footer-overlap.md
+      - 答辩总结页没有关闭按钮: bug_fix/voice-defense-ended-missing-close.md
+      - 答辩开场失败关闭无响应: bug_fix/voice-defense-close-before-playback.md
       - 远端 ACP PATH / SSH: bug_fix/remote-acp-path-ssh.md
       - 远程 Vault 文件树懒加载: bug_fix/remote-vault-tree-lazy.md
       - macOS 签名与公证: bug_fix/macos-signing.md
