+++ src-tauri/capabilities/viva-window.json
@@ -1,0 +1,7 @@
+{
+	"$schema": "../gen/schemas/desktop-schema.json",
+	"identifier": "viva-window",
+	"description": "Force-close permission required by Tauri onCloseRequested for the Viva window",
+	"windows": ["viva"],
+	"permissions": ["core:window:allow-destroy"]
+}
