+++ src/i18n/locales/en/agent.json
@@ -74,6 +74,9 @@
 			"completed": "Defense material is ready for review.",
 			"cancelFailed": "Could not cancel defense preparation.",
 			"saveBriefFailed": "Could not confirm defense material.",
+			"startUnavailable": "The defense material is still loading. Try again in a moment.",
+			"startBlocked": "A defense connection is already starting or active. Please wait.",
+			"startScopeChanged": "The defense preparation changed. Reopen the viva before entering.",
 			"completeFailed": "Could not update the preparation record.",
 			"stillRunning": "Wait for defense preparation to finish before entering the viva.",
 			"reviewRunning": "Wait for the defense review to finish before entering the viva.",
@@ -140,6 +143,7 @@
 			"microphoneDenied": "Microphone access was denied. Allow it in system settings.",
 			"microphoneNotFound": "No microphone was found.",
 			"microphoneBusy": "The microphone is in use by another application.",
+			"openingTimeout": "The committee did not begin with the required opening. Please try again.",
 			"dataChannelTimeout": "The realtime data connection timed out.",
 			"dataChannelFailed": "The realtime data connection failed.",
 			"networkOffline": "The network connection was lost.",
