+++ src/i18n/locales/zh-CN/agent.json
@@ -74,6 +74,9 @@
 			"completed": "答辩材料已生成，可以检查了。",
 			"cancelFailed": "无法取消答辩准备。",
 			"saveBriefFailed": "无法确认答辩材料。",
+			"startUnavailable": "答辩材料还在加载，请稍后再试。",
+			"startBlocked": "已有答辩连接正在启动或进行中，请稍候。",
+			"startScopeChanged": "答辩准备状态已变化，请重新打开答辩间后再入场。",
 			"completeFailed": "无法更新答辩准备记录。",
 			"stillRunning": "请等待答辩准备结束后再入场。",
 			"reviewRunning": "请等待答辩评价结束后再入场。",
@@ -140,6 +143,7 @@
 			"microphoneDenied": "麦克风权限被拒绝，请在系统设置中允许。",
 			"microphoneNotFound": "未找到可用的麦克风。",
 			"microphoneBusy": "麦克风正在被其他应用占用。",
+			"openingTimeout": "委员未按要求开始答辩，请重试。",
 			"dataChannelTimeout": "实时数据连接超时。",
 			"dataChannelFailed": "实时数据连接失败。",
 			"networkOffline": "网络连接已断开。",
