+++ src/lib/voice-defense/index.ts
@@ -5,9 +5,12 @@ export * from "@/lib/voice-defense/debrief";
 export * from "@/lib/voice-defense/errors";
 export * from "@/lib/voice-defense/metrics";
 export * from "@/lib/voice-defense/open-request";
+export * from "@/lib/voice-defense/preferences";
 export * from "@/lib/voice-defense/preparation";
+export * from "@/lib/voice-defense/preparation-recovery";
 export * from "@/lib/voice-defense/protocol";
 export * from "@/lib/voice-defense/review";
+export * from "@/lib/voice-defense/start-gate";
 export * from "@/lib/voice-defense/thought-stream";
 export * from "@/lib/voice-defense/timing";
 export * from "@/lib/voice-defense/transcript";
